Fixed: advanced section store file
